fixed health check issue
