+++ src/main/java/com/opuscapita/peppol/smp/controller/SmpHomeController.java
@@ -1,6 +1,5 @@
 package com.opuscapita.peppol.smp.controller;
 
-import com.google.gson.JsonObject;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 
@@ -10,12 +9,5 @@ public class SmpHomeController {
     @RequestMapping(value = "/")
     public String index() {
         return "index";
-    }
-
-    @RequestMapping("/api/health/check")
-    public JsonObject health() {
-        JsonObject response = new JsonObject();
-        response.addProperty("message", "Yes, I'm alive!");
-        return response;
     }
 }
